backend/requirements.txt
fastapi
python-jose
bcrypt
uvicorn

httpx
pytest
